app/requirements.txt
uvicorn==0.24.0.post1
fastapi==0.104.1
pydantic==2.5.2
typing==3.7.4.3
psycopg2-binary==2.9.9